+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 [unreleased]
+- Change: Added input validation to zprobe offset inputs
 - Fix: Sometimes the machine doesn't response to the initial machine "model" or "version" queries. Attempt to query this machine metadata periodically until it's determined
-- Fix: Fixed single axis z probing.
+- Fix: Fixed single axis z probing
 
 
 [0.10.0]